requirements.txt
langchain>=0.1.0
langchain-community>=0.0.1
googlemaps>=4.10.0
python-dotenv>=1.0.0
openai>=1.0.0